Fix zshrc
- Fix autocomplete key so it works with emacs terminal (C-SPC is bound
to something company mode). SHIFT-TAB is good because a) it's like the
existing tab autocomplete, but modified, and is compatible with emacs,
so win-win.

- Change remote host zsh theme so it's not confusing when I ssh into
something from my local. Rn set to "cloud" theme but we'll see what
sticks. In any case it definitely needs to be different than my "pc" theme.